# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для управління новинами з блоковою структурою контенту.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user_news": {
            "get": {
                "tags": [
                    "User News"
                ],
                "summary": "Отримати список власних новин",
                "operationId": "5e28ec22584c9eeeac673cff5c8370ca",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Пошук",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_published",
                        "in": "query",
                        "description": "Фільтр за статусом",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кількість на сторінку",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер сторінки",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список новин користувача"
                    },
                    "401": {
                        "description": "Неавторизований"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User News"
                ],
                "summary": "Створити новину",
                "operationId": "b1f57450a61284aa43ee78e67bc9ace7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Назва новини"
                                    },
                                    "short_description": {
                                        "type": "string",
                                        "example": "Короткий опис"
                                    },
                                    "image": {
                                        "description": "Картинка новини",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Новину створено"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user_news/{id}": {
            "get": {
                "tags": [
                    "User News"
                ],
                "summary": "Отримати конкретну новину користувача",
                "operationId": "dd5bf5509734f79c531f2600f3c5edb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новини",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Дані новини"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "404": {
                        "description": "Не знайдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User News"
                ],
                "summary": "Оновити новину (POST з _method=PUT для файлів)",
                "operationId": "f41183eca4cd12327fee0ee93f4fe46e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новини",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Спуфінг методу",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Оновлена назва"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_published": {
                                        "type": "boolean"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Новину оновлено"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User News"
                ],
                "summary": "Видалити новину",
                "operationId": "f67cca6e3439ba82f15dc418185a7870",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новини",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Новину видалено"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "404": {
                        "description": "Не знайдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Отримати профіль поточного користувача",
                "operationId": "7d12955ead9260d82af59043cdf7f84d",
                "responses": {
                    "200": {
                        "description": "Дані профілю"
                    },
                    "401": {
                        "description": "Неавторизований"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Оновити профіль (POST з _method=PUT для файлів)",
                "operationId": "4f10e8f1a9b5877a8fe852d886240504",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Спуфінг методу",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Новий Макс"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "bio": {
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "description": "Аватар",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Профіль оновлено"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введіть токен, отриманий при логіні",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}